+++ JavaWebServiceGroupB/src/main/java/com/kainos/ea/employee_stuff/SalesEmployee.java
@@ -7,7 +7,7 @@ public class SalesEmployee extends Employee {
     private float commissionRate;
     private int salesTotal;
 
-    public SalesEmployee(short number, int salary, String firstName, String lastName, String bankAccountNumber, String niNumber, String phoneNumber, String email, float commissionRate, int salesTotal) {
+    public SalesEmployee(short number, double salary, String firstName, String lastName, String bankAccountNumber, String niNumber, String phoneNumber, String email, float commissionRate, int salesTotal) {
         super(number, salary, firstName, lastName, bankAccountNumber, niNumber, phoneNumber, email);
         this.commissionRate = commissionRate;
         this.salesTotal = salesTotal;
@@ -30,7 +30,7 @@ public class SalesEmployee extends Employee {
     }
 
     @Override
-    public int calcPay() {
+    public double calcPay() {
         return super.calcPay() + Math.round(commissionRate * salesTotal);
     }
